+++ app/src/main/java/com/wintersky/windyreader/data/source/local/LocalDataSource.java
@@ -26,6 +26,7 @@ import com.wintersky.windyreader.util.AppExecutors;
 
 import java.io.BufferedInputStream;
 import java.io.BufferedOutputStream;
+import java.io.ByteArrayOutputStream;
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileOutputStream;
@@ -49,7 +50,7 @@ import io.realm.exceptions.RealmException;
 @Singleton
 public class LocalDataSource implements DataSource {
 
-    public static String CACHE_DIR = "BookCache/";
+    public static final String CACHE_DIR = "BookCache/";
     private final Context mContext;
     private final AppExecutors mExecutors;
     private final Realm mRealm;
@@ -85,14 +86,15 @@ public class LocalDataSource implements DataSource {
 
     @Override
     public void getCatalog(@NonNull final Book book, @NonNull final GetCatalogCallback callback) {
-        final Book bk = mRealm.where(Book.class).equalTo("catalogUrl", book.getCatalogUrl()).findFirstAsync();
+        final String ctUrl = book.getCatalogUrl();
+        final Book bk = mRealm.where(Book.class).equalTo("catalogUrl", ctUrl).findFirstAsync();
         bk.addChangeListener(new RealmObjectChangeListener<RealmModel>() {
             @Override
             public void onChange(@NonNull final RealmModel model, @Nullable final ObjectChangeSet changeSet) {
                 if (bk.isValid()) {
                     callback.onLoaded(bk.getCatalog());
                 } else {
-                    callback.onFailed(new RealmException("Book not find: (catalogUrl) " + book.getCatalogUrl()));
+                    callback.onFailed(new RealmException("Book not find: (catalogUrl) " + ctUrl));
                 }
                 bk.removeAllChangeListeners();
             }
@@ -123,6 +125,9 @@ public class LocalDataSource implements DataSource {
 
     @Override
     public void saveBook(@NonNull final Book book, @NonNull final SaveBookCallback callback) {
+        if (book.isManaged()) {
+            callback.onFailed(new RealmException("Book exist: " + book.getUrl()));
+        }
         mRealm.executeTransactionAsync(new Realm.Transaction() {
             @Override
             public void execute(@NonNull Realm realm) {
@@ -142,31 +147,25 @@ public class LocalDataSource implements DataSource {
     }
 
     @Override
-    public void deleteBook(@NonNull String bkUrl, @NonNull final DeleteBookCallback callback) {
-        getBook(bkUrl, new GetBookCallback() {
+    public void deleteBook(@NonNull final String bkUrl, @NonNull final DeleteBookCallback callback) {
+        mRealm.executeTransactionAsync(new Realm.Transaction() {
             @Override
-            public void onLoaded(@NonNull final Book book) {
-                mRealm.executeTransactionAsync(new Realm.Transaction() {
-                    @Override
-                    public void execute(@NonNull final Realm realm) {
-                        book.getCatalog().deleteAllFromRealm();
-                        book.deleteFromRealm();
-                    }
-                }, new Realm.Transaction.OnSuccess() {
-                    @Override
-                    public void onSuccess() {
-                        callback.onDeleted();
-                    }
-                }, new Realm.Transaction.OnError() {
-                    @Override
-                    public void onError(@NonNull final Throwable error) {
-                        callback.onFailed(error);
-                    }
-                });
+            public void execute(@NonNull final Realm realm) {
+                Book book = realm.where(Book.class).equalTo("url", bkUrl).findFirst();
+                if (book == null) {
+                    throw new RealmException("Book not find: " + bkUrl);
+                }
+                book.getCatalog().deleteAllFromRealm();
+                book.deleteFromRealm();
             }
-
+        }, new Realm.Transaction.OnSuccess() {
             @Override
-            public void onFailed(@NonNull final Throwable error) {
+            public void onSuccess() {
+                callback.onDeleted();
+            }
+        }, new Realm.Transaction.OnError() {
+            @Override
+            public void onError(@NonNull final Throwable error) {
                 callback.onFailed(error);
             }
         });
@@ -178,9 +177,12 @@ public class LocalDataSource implements DataSource {
             return false;
         }
         File bkDir = new File(root, CACHE_DIR + ctUrl.replace("/", "_"));
-        bkDir.mkdirs();
+        if (!bkDir.exists() && !bkDir.mkdirs()) {
+            return false;
+        }
         File file = new File(bkDir, chUrl.replace("/", "_"));
         try {
+            //noinspection ResultOfMethodCallIgnored
             file.createNewFile();
         } catch (IOException e) {
             e.printStackTrace();
@@ -190,9 +192,9 @@ public class LocalDataSource implements DataSource {
                 BufferedOutputStream bos = new BufferedOutputStream(fos)
         ) {
             bos.write(content.getBytes());
-            bos.flush();
         } catch (IOException e) {
             e.printStackTrace();
+            //noinspection ResultOfMethodCallIgnored
             file.delete();
             return false;
         }
@@ -210,15 +212,20 @@ public class LocalDataSource implements DataSource {
                 FileInputStream fis = new FileInputStream(file);
                 BufferedInputStream bis = new BufferedInputStream(fis)
         ) {
-            byte[] buff = new byte[bis.available()];
-            bis.read(buff);
-            return new String(buff);
+            byte[] buff = new byte[1024];
+            ByteArrayOutputStream ba = new ByteArrayOutputStream();
+            int len;
+            while ((len = bis.read(buff)) != -1) {
+                ba.write(buff, 0, len);
+            }
+            return ba.toString();
         } catch (IOException e) {
             e.printStackTrace();
             return null;
         }
     }
 
+    @SuppressWarnings("BooleanMethodIsAlwaysInverted")
     public boolean isContentExist(@NonNull String ctUrl, @NonNull String chUrl) {
         File root = mContext.getExternalFilesDir("");
         if (root == null) {
